fix(dns,io): Android/ARM32 cross-compilation issues
Multiple cross-platform compilation fixes for Android and ARM32:

* dap_context.c:
  - rlim_t is unsigned long on Android (32-bit), not uint64_t
  - Added explicit casts to uint64_t for DAP_UINT64_FORMAT_U

* dap_dns_server.h:
  - Changed single-bit bitfield types from 'int' to 'unsigned int'
  - Clang treats assignment 1 to signed 1-bit field as truncation (-1)
  - Affects: qr, aa, tc, rd, ra, z, rcode, opcode fields

* scripts/test-android-local.sh:
  - New local Android ARMv7 NEON cross-compilation test script
  - Uses Android NDK 26.1.10909125 with CMake toolchain
  - Full build verification without device deployment

Root cause:
- Single-bit signed int bitfields can only hold 0 or -1
- Clang warns: "implicit truncation from 'int' to -1"
- unsigned int bitfields correctly hold 0 or 1

Tested:
- ✅ ARM32 (armv7 + NEON): PASS
- ✅ Android (armv7-a): PASS
- ✅ x86_64 Linux: PASS

+++ module/io/dap_context.c
@@ -92,7 +92,8 @@ int dap_context_init()
     l_fdlimit.rlim_cur = l_fdlimit.rlim_max;
     if (setrlimit(RLIMIT_NOFILE, &l_fdlimit))
         return -2;
-    log_it(L_INFO, "Set maximum opened descriptors from %" DAP_UINT64_FORMAT_U " to %"DAP_UINT64_FORMAT_U, l_oldlimit, l_fdlimit.rlim_cur);
+    log_it(L_INFO, "Set maximum opened descriptors from %" DAP_UINT64_FORMAT_U " to %"DAP_UINT64_FORMAT_U, 
+           (uint64_t)l_oldlimit, (uint64_t)l_fdlimit.rlim_cur);
 #endif
     return 0;
 }
